requirements.txt
albumentations==1.3.0
coremltools==6.1
ipython==8.7.0
matplotlib==3.6.2
numpy==1.24.0
onnx==1.13.0
onnx_graphsurgeon==0.0.1.dev5
onnxsim==0.4.10
opencv_python_headless==4.6.0.66
pafy==0.5.5
pandas==1.5.2
Pillow==9.3.0
pycocotools==2.0.6
PyYAML==6.0
requests==2.28.1
scikit_learn==1.2.0
scipy==1.7.3
seaborn==0.12.1
setuptools==65.6.3
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
tritonclient==2.29.0
wandb==0.13.7
